refactor: 字符管线核心 wchar_t → char32_t（根治代理对拆分）

+++ lib/src/util/History.h
@@ -210,7 +210,7 @@ public:
 
     CharacterColor fgColor, bgColor;
     quint16 startPos;
-    quint8 rendition;
+    quint16 rendition;
 };
 
 class HistoryType
